[State-Integrator] First draft of the integrator
To be plugged before the device if integration of the control is needed:
Only for control in Velocity and Acceleration for now.
Check the size of the signal to separate two cases: with and without a control on the freeflyer.
The integration of the freeflyer works as an odometry predictive system.