[Device] Add a unit test for state integrator
Use device.cpp from sot-core. Add yaml files to create the device for the test.
Set the input control of the integrator to -0.5 m/s (velocity) for every joints.
Plug the output stateSOUT_ of the integrator to the input control of the device. The control type of the joints on the hardware side (in device) is in position.
At the end of the loop the output of the device (motor control) correspond to the limits of the joints in position (as expected, the velocity control push the joints to their limits).
In the test a freeflyer is given and integrated as an odometry system.