[State-integrator] Separate integrate function in 2: for joints or freeflyer
Add entry signal for freeflyer (freeFlyerSIN), add a variable to define controlType of the Freeflyer (controlTypeFF_) with setter. 
Add two output signals for freeflyer one for Euler angles (freeFlyerPositionEulerSOUT_) and the other for quaternions (freeFlyerPositionQuatSOUT_).
Change the computation of the Euler angles to assure their range to be [-pi:pi]x[-pi/2:pi/2]x[-pi:pi].